Makefile: allow several "make install" for update
For example:
<shell>
$ make install PREFIX=/usr
...
... install files ...
...
$ make install PREFIX=/usr
$
</shell>

The second make did nothing. This is not the case anymore.
This is a common user behavior for "update".

Alternate clean way is:
<shell>
$ make install PREFIX=/usr
$ make uninstall PREFIX=/usr
$ make install PREFIX=/usr
</shell>

Note: "update" and "reinstall" are a little different.
In first case, old files (dropped modules, orphaned files) are kept, but this is not so important.